% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{bm25_tokens_list}
\alias{bm25_tokens_list}
\title{Convert input list of raw tokens to a list of tokens and corresponding
frequencies.}
\usage{
bm25_tokens_list(txt)
}
\arguments{
\item{txt}{An optional list of input documents. If not specified, data will
be loaded as specified by the \code{corpus} parameter.}
}
\value{
A list of \code{data.frame} objects, one for each input item, and each
including two columns of "token" and "n" holding frequencies for each token.
}
\description{
Convert input list of raw tokens to a list of tokens and corresponding
frequencies.
}
