% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetreesim-package.R
\docType{package}
\name{codetreesim-package}
\alias{codetreesim}
\alias{codetreesim-package}
\title{codetreesim: Tree Similarity Metrics for R Code}
\description{
Tree similarity metrics for R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/codetreesim/}
  \item \url{https://github.com/ropensci-review-tools/codetreesim}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/codetreesim/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Other contributors:
\itemize{
  \item Mateusz Pawlik (Author of original tree-similarity src code) [copyright holder]
  \item Nikolaus Augsten (Author of original tree-similarity src code) [copyright holder]
  \item Daniel Kocher (Author of original tree-similarity src code) [copyright holder]
  \item Thomas Huetter (Author of original tree-similarity src code) [copyright holder]
}

}
\keyword{internal}
