% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{head.pkgmatch}
\alias{head.pkgmatch}
\title{Head method for 'pkgmatch' objects}
\usage{
\method{head}{pkgmatch}(x, n = 5L, ...)
}
\arguments{
\item{x}{Object for which head is to be printed}

\item{n}{Number of rows of full \code{pkgmatch} object to be displayed}

\item{...}{Not used}
}
\description{
Head method for 'pkgmatch' objects
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
p <- pkgmatch_similar_pkgs (input)
p # Default print method, lists 5 best matching packages
head (p) # Shows first 5 rows of full `data.frame` object
}
}
\seealso{
Other utils: 
\code{\link{pkgmatch_load_data}()},
\code{\link{print.pkgmatch}()},
\code{\link{text_is_code}()}
}
\concept{utils}
