% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-package.R
\docType{package}
\name{pkgmatch-package}
\alias{pkgmatch}
\alias{pkgmatch-package}
\title{pkgmatch: Find R Packages Matching Either Descriptions or Other R Packages}
\description{
Find R packages matching either descriptions or other R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgmatch/}
  \item \url{https://github.com/ropensci-review-tools/pkgmatch}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgmatch/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Other contributors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co} [contributor]
}

}
\keyword{internal}
