% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgmatch_bm25}
\alias{pkgmatch_bm25}
\title{Calculate the "BM25" = "Best Matching 25" ranking function between text
input and all R packages within specified corpus.}
\usage{
pkgmatch_bm25(input, txt = NULL, idfs = NULL, corpus = "ropensci")
}
\arguments{
\item{input}{A single character string to match against the second parameter
of all input documents.}

\item{txt}{An optional list of input documents. If not specified, data will
be loaded as specified by the \code{corpus} parameter.}

\item{idfs}{Optional list of Inverse Document Frequency weightings generated
by the internal \code{bm25_idf} function. If not specified, values for the
rOpenSci corpus will be automatically downloaded and used.}

\item{corpus}{If \code{txt} is not specified, data for nominated corpus will be
downloaded to local cache directory, and BM25 values calculated against
those. Must be one of "ropensci", "ropensci-fns".}
}
\value{
A \code{data.frame} of package names and 'BM25' measures against text
from whole packages both with and without function descriptions.
}
\description{
See \url{https://en.wikipedia.org/wiki/Okapi_BM25}.
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
bm25 <- pkgmatch_bm25 (input)
# Or pre-load document-frequency weightings:
idfs <- pkgmatch_load_data ("idfs", fns = FALSE)
bm25 <- pkgmatch_bm25 (input, idfs = idfs)
}
}
\seealso{
Other bm25: 
\code{\link{pkgmatch_bm25_fn_calls}()}
}
\concept{bm25}
