% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgmatch_embeddings_from_text}
\alias{pkgmatch_embeddings_from_text}
\title{Return raw 'LLM' embeddings from a vector of text strings.}
\usage{
pkgmatch_embeddings_from_text(input = NULL)
}
\arguments{
\item{input}{A vector of one or more text strings for which embeddings are
to be extracted.}
}
\value{
A matrix of embeddings, one column for each \code{input} item, and a
fixed number of rows defined by the embedding length of the language models.
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
emb <- pkgmatch_embeddings_from_text (input = input)
}
}
\seealso{
Other embeddings: 
\code{\link{pkgmatch_embeddings_from_pkgs}()}
}
\concept{embeddings}
