% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{pkgmatch_treesitter_fn_tags}
\alias{pkgmatch_treesitter_fn_tags}
\title{Use "treesitter" to tag all function calls made within local package, and to
associate those calls with package namespaces. This is used as input to the
\link{pkgmatch_bm25_fn_calls} function.}
\usage{
pkgmatch_treesitter_fn_tags(path)
}
\arguments{
\item{path}{Path to local package, or \code{.tar.gz} file of package source.}
}
\value{
A \code{data.frame} of all function calls made within the package, with
the following columns:
\itemize{
\item 'fn' Name of the package function within which call is made, including
namespace identifiers of "::" for exported functions and ":::" for
non-exported functions.
\item name Name of function being called, including namespace.
\item start Byte number within file corresponding to start of definition
\item end Byte number within file corresponding to end of definition
\item file Name of file in which fn call is defined.
}
}
\description{
Use "treesitter" to tag all function calls made within local package, and to
associate those calls with package namespaces. This is used as input to the
\link{pkgmatch_bm25_fn_calls} function.
}
\examples{
\dontrun{
u <- "https://cran.r-project.org/src/contrib/odbc_1.5.0.tar.gz"
path <- file.path (tempdir (), basename (u))
download.file (u, destfile = path)
tags <- pkgmatch_treesitter_fn_tags (path)
}
}
\concept{treesitter}
