% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsimil-package.R
\docType{package}
\name{pkgsimil-package}
\alias{pkgsimil}
\alias{pkgsimil-package}
\title{pkgsimil: Similarity Metrics between R Packages}
\description{
Similarity metrics between R packages, based on similarity of code and documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgsimil/}
  \item \url{https://github.com/ropensci-review-tools/pkgsimil}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgsimil/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
