% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgsimil_bm25}
\alias{pkgsimil_bm25}
\title{Calculate the "BM25" = "Best Matching 25" ranking function.}
\usage{
pkgsimil_bm25(input, txt)
}
\arguments{
\item{input}{A single character string to match against the second parameter
of all input documents.}

\item{txt}{A list of input documents.}
}
\description{
See \url{https://en.wikipedia.org/wiki/Okapi_BM25}.
}
