% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding-similarity-desc.R
\name{pkgsimil_embed_desc}
\alias{pkgsimil_embed_desc}
\title{Calculate distances between embeddings from 'Description' entries  of packages.}
\usage{
pkgsimil_embed_desc(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{Name of one or more packages for which trees are to be extracted.}
}
\description{
Calculate distances between embeddings from 'Description' entries  of packages.
}
