% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding-similarity-desc.R
\name{pkgsimil_embed_fns}
\alias{pkgsimil_embed_fns}
\title{Calculate minimal distances between embeddings from function descriptions.}
\usage{
pkgsimil_embed_fns(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{Name of one or more packages for which trees are to be extracted.}
}
\description{
The embeddings are currently retrieved from the Jina AI API. A distance is
calculated for each function in each package, as the minimal distance to all
functions of all other packages. The distance between two packages is then
the average across all functions of those minimal distances.
}
