% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding-similarity-desc.R
\name{pkgsimil_embeddings}
\alias{pkgsimil_embeddings}
\title{Calculate distances between 'LLM' embeddings from package text and function
definitions.}
\usage{
pkgsimil_embeddings(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{Name of one or more packages for which trees are to be extracted.}
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
