% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgsimil_embeddings_from_text}
\alias{pkgsimil_embeddings_from_text}
\title{Return raw 'LLM' embeddings from a vector of text strings.}
\usage{
pkgsimil_embeddings_from_text(input = NULL)
}
\arguments{
\item{input}{A vector of one or more text strings for which embeddings are
to be extracted.}
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
