% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgsimil_load_embeddings}
\alias{pkgsimil_load_embeddings}
\title{Load embeddings generated by the \link{pkgsimil_embeddings_raw} function,
either for all rOpenSci packages or, if \code{fns = TRUE},  all individual
functions within those packages.}
\usage{
pkgsimil_load_embeddings(fns = FALSE)
}
\arguments{
\item{fns}{If \code{FALSE} (default), load embeddings for all rOpenSci packages;
otherwise load (considerably larger dataset of) embeddings for all
individual functions.}
}
\description{
Load embeddings generated by the \link{pkgsimil_embeddings_raw} function,
either for all rOpenSci packages or, if \code{fns = TRUE},  all individual
functions within those packages.
}
