% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{print.pkgmatch}
\alias{print.pkgmatch}
\title{Print method for 'pkgmatch' objects}
\usage{
\method{print}{pkgmatch}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Not used}
}
\description{
Print method for 'pkgmatch' objects
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
p <- pkgmatch_similar_pkgs (input)
p # Default print method, lists 5 best matching packages
head (p) # Shows first 5 rows of full `data.frame` object
}
}
\seealso{
Other utils: 
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{text_is_code}()}
}
\concept{utils}
