% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsimil-methods.R
\name{print.pkgsimil}
\alias{print.pkgsimil}
\title{Print method for 'pkgsimil' objects}
\usage{
\method{print}{pkgsimil}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Not used}
}
\description{
Print method for 'pkgsimil' objects
}
\seealso{
Other utils: 
\code{\link{pkgsimil_load_data}()},
\code{\link{text_is_code}()}
}
\concept{utils}
