% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-get.R
\name{tree_get}
\alias{tree_get}
\title{Get a syntax tree for a specified package, including all functions both
exported and internal.}
\usage{
tree_get(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{Name of one or more packages for which trees are to be extracted.}
}
\value{
A character vector of syntax trees, one for each nominated package.
}
\description{
Get a syntax tree for a specified package, including all functions both
exported and internal.
}
\examples{
\dontrun{
trees <- tree_get (c ("utils", "tools"))
# These are huge syntax trees:
sapply (trees, function (i) format (nchar (i), big.mark = ","))
}
}
