% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-similarity.R
\name{tree_similarity}
\alias{tree_similarity}
\title{Tree similarity}
\usage{
tree_similarity(trees)
}
\arguments{
\item{trees}{Input trees as returned from \link{tree_parse} function, either
as list or vector of character objects with each element holding one tree.}
}
\value{
Similarity metrics
}
\description{
Tree similarity
}
