% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesim.R
\name{treesim}
\alias{treesim}
\title{treesim function'}
\usage{
treesim(node_brackets = TRUE)
}
\arguments{
\item{node_brackets}{If \code{TRUE} (default), wrap all nodes in S-expression braces.}
}
\description{
treesim function'
}
