% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgmatch_bm25_fn_calls}
\alias{pkgmatch_bm25_fn_calls}
\title{Calculate a "BM25" index from function-call frequencies between a local R
package and all packages in specified corpus.}
\usage{
pkgmatch_bm25_fn_calls(path, corpus = "ropensci")
}
\arguments{
\item{path}{Local path to source code of an R package.}

\item{corpus}{One of "ropensci" or "cran"}
}
\description{
Calculate a "BM25" index from function-call frequencies between a local R
package and all packages in specified corpus.
}
\examples{
\dontrun{
u <- "https://cran.r-project.org/src/contrib/odbc_1.5.0.tar.gz"
path <- file.path (tempdir (), basename (u))
download.file (u, destfile = path)
bm25 <- pkgmatch_bm25_fn_calls (path)
}
}
\seealso{
Other bm25: 
\code{\link{pkgmatch_bm25}()}
}
\concept{bm25}
