% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgsimil_bm25}
\alias{pkgsimil_bm25}
\title{Calculate the "BM25" = "Best Matching 25" ranking function.}
\usage{
pkgsimil_bm25(input, txt = NULL, idfs = NULL, corpus = "ropensci")
}
\arguments{
\item{input}{A single character string to match against the second parameter
of all input documents.}

\item{txt}{An optional list of input documents. If not specified, data will
be loaded as specified by the \code{corpus} parameter.}

\item{idfs}{Optional list of Inverse Document Frequency weightings generated
by the internal \code{bm25_idf} function. If not specified, values for the
rOpenSci corpus will be automatically downloaded and used.}

\item{corpus}{If \code{txt} is not specified, data for nominated corpus will be
downloaded to local cache directory, and BM25 values calculated against
those. Must be one of "ropensci", "ropensci-fns".}
}
\value{
A \code{data.frame} of package names and 'BM25' measures against text
from whole packages both with and without function descriptions.
}
\description{
See \url{https://en.wikipedia.org/wiki/Okapi_BM25}.
}
