% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgsimil_bm25_fn_calls}
\alias{pkgsimil_bm25_fn_calls}
\title{Calculate a "BM25" index from function-call frequencies between a local R
package and all packages in specified corpus.}
\usage{
pkgsimil_bm25_fn_calls(path, corpus = "ropensci")
}
\arguments{
\item{path}{Local path to source code of an R package.}

\item{corpus}{One of "ropensci" or "cran"}
}
\description{
Calculate a "BM25" index from function-call frequencies between a local R
package and all packages in specified corpus.
}
\seealso{
Other bm25: 
\code{\link{pkgsimil_bm25}()}
}
\concept{bm25}
