% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-embeddings.R
\name{pkgsimil_embeddings}
\alias{pkgsimil_embeddings}
\title{Calculate distances between 'LLM' embeddings from package text and function
definitions.}
\usage{
pkgsimil_embeddings(packages = NULL)
}
\arguments{
\item{packages}{Names of, or paths to,  one or more packages for which
embedding similarities are to be calculated.}
}
\value{
A \code{data.frame} of pair-wise similarities between all packages
specified in \code{packages}.
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
