% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgsimil_embeddings_from_pkgs}
\alias{pkgsimil_embeddings_from_pkgs}
\title{Return raw 'LLM' embeddings from package text and function definitions.}
\usage{
pkgsimil_embeddings_from_pkgs(packages = NULL, functions_only = FALSE)
}
\arguments{
\item{packages}{A vector of local paths to directories containing R packages.}

\item{functions_only}{If \code{TRUE}, calculate embeddings for function
descriptions only. This is intended to generate a separate set of embeddings
which can then be used to match plain-text queries of functions, rather than
entire packages.}
}
\value{
If \code{!functions_only}, a list of two matrices of embeddings: one for
the text descriptions of the specified packages, including individual
descriptions of all package functions, and one for the entire code base. For
\code{functions_only}, a single matrix of embeddings for all function
descriptions.
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
\examples{
\dontrun{
packages <- c ("cli", "fs")
emb_fns <- pkgsimil_embeddings_from_pkgs (packages, functions_only = TRUE)
colnames (emb_fns) # All functions of the two packages
emb_pkg <- pkgsimil_embeddings_from_pkgs (packages, functions_only = FALSE)
names (emb_pkg) # text_with_fns, text_wo_fns, code
colnames (emb_pkg$text_with_fns) # cli, fs
}
}
\seealso{
Other embeddings: 
\code{\link{pkgsimil_embeddings_from_text}()}
}
\concept{embeddings}
