% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgsimil_embeddings_raw}
\alias{pkgsimil_embeddings_raw}
\title{Return raw 'LLM' embeddings from package text and function definitions.}
\usage{
pkgsimil_embeddings_raw(packages = NULL)
}
\arguments{
\item{packages}{Names of, or paths to,  one or more packages for which
embedding similarities are to be calculated.}
}
\value{
A list of two matrices of embeddings: one for the text descriptions
of the specified packages, including individual descriptions of all package
functions, and one for the entire code base.
}
\description{
The embeddings are currently retrieved from a local 'ollama' server running
Jina AI embeddings.
}
