% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pkgsimil_load_data}
\alias{pkgsimil_load_data}
\title{Load embeddings generated by the \link{pkgsimil_embeddings_from_pkgs}
function, either for all rOpenSci packages or, if \code{fns = TRUE},  all
individual functions within those packages.}
\usage{
pkgsimil_load_data(what = "embeddings", fns = FALSE)
}
\arguments{
\item{what}{Either "embeddings" to load pre-generated embeddings, or "idfs"
to load pre-generated Inverse Document Frequency weightings.}

\item{fns}{If \code{FALSE} (default), load embeddings for all rOpenSci packages;
otherwise load (considerably larger dataset of) embeddings for all
individual functions.}
}
\value{
The loaded \code{data.frame}.
}
\description{
Load embeddings generated by the \link{pkgsimil_embeddings_from_pkgs}
function, either for all rOpenSci packages or, if \code{fns = TRUE},  all
individual functions within those packages.
}
\examples{
\dontrun{
embeddings <- pkgsimil_load_data ("embeddings")
embeddings_fns <- pkgsimil_load_data ("embeddings", fns = TRUE)
idfs <- pkgsimil_load_data ("idfs")
idfs_fns <- pkgsimil_load_data ("idfs", fns = TRUE)
}
}
