% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-fns.R
\name{pkgsimil_similar_fns}
\alias{pkgsimil_similar_fns}
\title{Identify functions best matching a given input string. Only applies to
functions from the corpus of rOpenSci packages.}
\usage{
pkgsimil_similar_fns(input, embeddings = NULL, n = 5L)
}
\arguments{
\item{input}{A text string.}

\item{embeddings}{A single matrix of embeddings produced from
\link{pkgsimil_embeddings_from_pkgs} with \code{functions_only = TRUE}. If not
cache directory.
provided, pre-generated embeddings will be downloaded and stored in a local}
}
\value{
A character vector of function names in the form
"\if{html}{\out{<package>}}::\if{html}{\out{<function>}}".
}
\description{
Identify functions best matching a given input string. Only applies to
functions from the corpus of rOpenSci packages.
}
\examples{
\dontrun{
input <- "Process raster satellite images"
pkgsimil_similar_fns (input)
}
}
\seealso{
Other main: 
\code{\link{pkgsimil_similar_pkgs}()}
}
\concept{main}
