% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-pkgs.R
\name{pkgsimil_similar_pkgs}
\alias{pkgsimil_similar_pkgs}
\title{Use the embeddings from \link{pkgsimil_embeddings_from_pkgs} to identify
most similar packages to a given local repository.}
\usage{
pkgsimil_similar_pkgs(
  input,
  embeddings = NULL,
  idfs = NULL,
  input_is_code = text_is_code(input),
  n = 5L
)
}
\arguments{
\item{input}{Either a path to local source code of an R package, or a text
string.}

\item{embeddings}{Large Language Model embeddings for all rOpenSci packages,
generated from \link{pkgsimil_embeddings_from_pkgs}. If not provided,
pre-generated embeddings will be downloaded and stored in a local cache
directory.}

\item{idfs}{Inverse Document Frequency tables for all rOpenSci packages,
generated from \link{pkgsimil_bm25}. If not provided, pre-generated IDF
tables will be downloaded and stored in a local cache directory.}

\item{input_is_code}{A binary flag indicating whether \code{input} is code or
plain text. Ignored if \code{input} is path to a local package; otherwise can be
used to force appropriate interpretation if input type.}

\item{n}{identify the \code{n} most similar packages in terms of both code and
text embeddings.}
}
\value{
If \code{input} is a path to a local package, a list of two character
vectors naming the \code{n} most similar packages in terms of descriptive textual
similarity ("text"), and in terms of similarity of code structure ("code").
If \code{input} is a single text string, a single character vector is returned
naming the \code{n} most similar packages.
}
\description{
Use the embeddings from \link{pkgsimil_embeddings_from_pkgs} to identify
most similar packages to a given local repository.
}
\seealso{
input_is_code
}
