% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest-pkgs(2).R, R/closest-pkgs.R
\name{pkgsimil_similar_pkgs}
\alias{pkgsimil_similar_pkgs}
\title{Use the embeddings from \link{pkgsimil_embeddings_raw} to identify most
similar packages to a given local repository.}
\usage{
pkgsimil_similar_pkgs(path, embeddings, n = 5L)

pkgsimil_similar_pkgs(path, embeddings, n = 5L)
}
\arguments{
\item{path}{Path to local repository}

\item{embeddings}{Large Language Model embeddings for all rOpenSci packages,
generated from \link{pkgsimil_embeddings_raw}.}

\item{n}{identify the \code{n} most similar packages in terms of both code and
text embeddings.}
}
\value{
A \code{data.frame} of all packages in the embeddings data and
corresponding distances.

A \code{data.frame} of all packages in the embeddings data and
corresponding distances.
}
\description{
Use the embeddings from \link{pkgsimil_embeddings_raw} to identify most
similar packages to a given local repository.

Use the embeddings from \link{pkgsimil_embeddings_raw} to identify most
similar packages to a given local repository.
}
