% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-pkgs.R
\name{pkgsimil_similar_pkgs}
\alias{pkgsimil_similar_pkgs}
\title{For an input of either a text string or local path to an R package, identify
most similar R packages from a specified corpus.}
\usage{
pkgsimil_similar_pkgs(
  input,
  corpus = "ropensci",
  embeddings = NULL,
  idfs = NULL,
  input_is_code = text_is_code(input),
  n = 5L
)
}
\arguments{
\item{input}{Either a path to local source code of an R package, or a text
string.}

\item{corpus}{If \code{embeddings} or \code{idfs} parameters are not specified, they
are automatically downloaded for the corpus specified by this parameter.
Must be one of "ropensci" or "cran". The function will then return the most
similar package from the specified corpus.}

\item{embeddings}{Large Language Model embeddings for all rOpenSci packages,
generated from \link{pkgmatch_embeddings_from_pkgs}. If not provided,
pre-generated embeddings will be downloaded and stored in a local cache
directory.}

\item{idfs}{Inverse Document Frequency tables for all rOpenSci packages,
generated from \link{pkgmatch_bm25}. If not provided, pre-generated IDF
tables will be downloaded and stored in a local cache directory.}

\item{input_is_code}{A binary flag indicating whether \code{input} is code or
plain text. Ignored if \code{input} is path to a local package; otherwise can be
used to force appropriate interpretation if input type.}

\item{n}{When the result of this function is printed to screen, the top \code{n}
packages will be displayed.}
}
\value{
If \code{input} is a path to a local package, a list of two \code{data.frame}
objects with data quantifying similarities in terms of descriptive textual
similarity ("text"), and in terms of similarity of code structure ("code").

vectors naming the \code{n} most similar packages in terms of descriptive textual
similarity ("text"), and in terms of similarity of code structure ("code").
If \code{input} is a single text string, a single \code{data.frame} object is returned
}
\description{
For an input of either a text string or local path to an R package, identify
most similar R packages from a specified corpus.
}
\note{
The first time this function is run without passing either
\code{embeddings} or \code{idfs}, required values will be automatically downloaded and
stored in a locally persistent cache directory. Especially for the "cran"
corpus, this downloading may take quite some time.
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
pkgsimil_similar_pkgs (input)
}
}
\seealso{
input_is_code

Other main: 
\code{\link{pkgsimil_similar_fns}()}
}
\concept{main}
