% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{print.pkgmatch}
\alias{print.pkgmatch}
\title{Print method for 'pkgmatch' objects}
\usage{
\method{print}{pkgmatch}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Not used}
}
\description{
Print method for 'pkgmatch' objects
}
\seealso{
Other utils: 
\code{\link{pkgmatch_load_data}()},
\code{\link{text_is_code}()}
}
\concept{utils}
