% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-similarity.R
\name{tree_similarity}
\alias{tree_similarity}
\title{Tree similarity}
\usage{
tree_similarity(trees, num_cores = 1L)
}
\arguments{
\item{trees}{Input trees as returned from \link{tree_parse} function, either
as list or vector of character objects with each element holding one tree.}

\item{num_cores}{Number of machine cores to use in parallel, defaulting to
single-core processing.}
}
\value{
A \code{data.frame} with each row containing pairwise similiarity metrics
between trees.
}
\description{
Tree similarity
}
