% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-similarity.R
\name{tree_similarity}
\alias{tree_similarity}
\title{Tree similarity}
\usage{
tree_similarity(tree)
}
\arguments{
\item{tree}{Input tree as S-expression.}
}
\value{
Similarity metrics
}
\description{
Tree similarity
}
